+++ projects/ui-lib/src/lib/services/mapper/component-mappers/index.ts
@@ -2,6 +2,8 @@ import { mapCategoryProgress } from './category-progress.mapper';
 import { mapHeaderClear } from './header-clear.mapper';
 import { mapHeroSection } from './hero-section.mapper';
 import { mapLangModal } from './lang-modal.mapper';
+import { mapLegalFooter } from './legal-footer.mapper';
+import { mapLinksFooter } from './links-footer.mapper';
 import { mapNavModal } from './nav-modal.mapper';
 import { mapSectionIntro } from './section-intro.mapper';
 
@@ -12,4 +14,6 @@ export const componentMappers: Record<string, (props: any) => any> = {
     'lang-modal': mapLangModal,
     'section-intro': mapSectionIntro,
     'category-progress': mapCategoryProgress,
+    'links-footer': mapLinksFooter,
+    'legal-footer': mapLegalFooter,
 };